[UHYU-289] refactor: benefit, mymap 폴더 정리
