--- src/features/mymap/components/mymap-list/index.ts
@@ -1,2 +1,0 @@
-export { default as AddMyMapButton } from './AddMyMapButton';
-export { default as MyMapList } from './MyMapList';
