+++ src/pages/mymap/MymapPage.tsx
@@ -1,11 +1,10 @@
-import MyMapBody from '@mymap/components/MymapBody';
-import { MymapHeader } from '@mymap/components/MymapHeader';
+import { MymapBody, MymapHeader } from '@mymap/components';
 
 const MymapPage = () => {
   return (
     <div className="flex flex-col gap-4 pb-10">
       <MymapHeader />
-      <MyMapBody />
+      <MymapBody />
     </div>
   );
 };
